# docs/api/api-docs.v1.json
{
  "components": {
    "schemas": {
      "OrderItemRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "상품 ID",
            "format": "int64",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "description": "주문 수량",
            "format": "int32",
            "example": 2
          }
        },
        "description": "주문 상품 정보"
      },
      "PostOrderRequestBody": {
        "type": "object",
        "properties": {
          "orderItems": {
            "type": "array",
            "description": "주문 상품 목록",
            "items": {
              "$ref": "#/components/schemas/OrderItemRequest"
            }
          },
          "couponId": {
            "type": "integer",
            "description": "사용할 쿠폰 ID (선택적)",
            "format": "int64",
            "example": 1
          }
        },
        "description": "주문 생성 요청"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.0.1",
  "paths": {
    "/api/v1/orders": {
      "post": {
        "tags": [
          "Order API"
        ],
        "summary": "주문 생성",
        "description": "상품을 주문하고 결제를 처리합니다.",
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostOrderRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "주문 생성 성공",
            "content": {
              "*/*": {
                "examples": {
                  "주문 생성 성공 응답": {
                    "description": "주문 생성 성공 응답",
                    "value": {
                      "orderId": 12345,
                      "totalAmount": 15000,
                      "usedPoint": 15000,
                      "couponDiscountAmount": 2000,
                      "orderItems": [
                        {
                          "productId": 1,
                          "productName": "아메리카노",
                          "unitPrice": 4500,
                          "quantity": 2,
                          "totalPrice": 9000
                        },
                        {
                          "productId": 2,
                          "productName": "라떼",
                          "unitPrice": 5000,
                          "quantity": 1,
                          "totalPrice": 5000
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않습니다.",
            "content": {
              "*/*": {
                "examples": {
                  "회원 없음 오류": {
                    "description": "회원 없음 오류",
                    "value": {
                      "message": "회원이 존재하지 않습니다."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "주문 처리 불가 (재고 부족, 포인트 부족, 쿠폰 사용 불가)",
            "content": {
              "*/*": {
                "examples": {
                  "재고 부족 오류": {
                    "description": "재고 부족 오류",
                    "value": {
                      "message": "상품 재고가 부족합니다."
                    }
                  },
                  "포인트 부족 오류": {
                    "description": "포인트 부족 오류",
                    "value": {
                      "message": "포인트 잔액이 부족합니다."
                    }
                  },
                  "쿠폰 사용 불가 오류": {
                    "description": "쿠폰 사용 불가 오류",
                    "value": {
                      "message": "쿠폰을 사용할 수 없습니다."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "주문 처리 실패",
            "content": {
              "*/*": {
                "examples": {
                  "시스템 오류": {
                    "description": "시스템 오류",
                    "value": {
                      "message": "주문 처리 중 오류가 발생했습니다."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
}